# typescript/services/control-plane-api/src/app.ts
import Fastify, { type FastifyInstance, type FastifyReply } from "fastify";
import {
  CipControlPlane,
  CipControlPlaneError,
  type CipEventBatch,
  type CipRepositories,
  type CompleteRunSessionInput,
  type RequestHumanApprovalInput,
  type ResolveApprovalRequestInput,
  type RollbackDeploymentInput,
  type StartRunSessionInput,
  type TransitionDeploymentInput,
} from "@new-odyssey/cip";

import {
  authenticateSdkApiKey,
  requireOperatorScope,
  type OperatorAuthConfig,
  verifyOperatorToken,
} from "./auth.js";
import {
  createQueuedIngestJob,
  hashPayload,
} from "./store.js";
import type {
  ApiKeyRecord,
  ApiKeyScope,
  ControlPlaneServiceStore,
  StoredHttpResponse,
} from "./types.js";
import {
  RequestValidationError,
  approvalRequestPathSchema,
  cipEventBatchSchema,
  completeRunSessionInputSchema,
  deploymentPathSchema,
  optionalTenantQuerySchema,
  parseOrThrow,
  requestHumanApprovalInputSchema,
  resolveApprovalRequestInputSchema,
  rollbackDeploymentInputSchema,
  sessionPathSchema,
  startRunSessionInputSchema,
  tenantPathSchema,
  transitionDeploymentInputSchema,
} from "./validation.js";

export interface ControlPlaneApiAppOptions {
  controlPlane: CipControlPlane;
  repositories: CipRepositories;
  serviceStore: ControlPlaneServiceStore;
  operatorAuth: OperatorAuthConfig;
  readyCheck?: () => Promise<boolean>;
}

const headerValue = (
  value: string | string[] | undefined,
): string | undefined => {
  if (Array.isArray(value)) {
    return value[0];
  }
  return value;
};

const requireIdempotencyKey = (headers: Record<string, unknown>): string => {
  const value = headerValue(headers["idempotency-key"] as string | string[] | undefined);
  if (value === undefined || value.trim() === "") {
    throw new Error("Idempotency-Key header is required");
  }
  return value;
};

const sendStoredResponse = async (
  response: StoredHttpResponse,
  reply: FastifyReply,
): Promise<void> => {
  await reply.code(response.status).send(response.body);
};

const routeKey = (method: string, path: string): string => `${method} ${path}`;

const sleep = async (durationMs: number): Promise<void> =>
  new Promise((resolve) => {
    setTimeout(resolve, durationMs);
  });

const waitForCompletedIdempotency = async (
  store: ControlPlaneServiceStore,
  route: string,
  idempotencyKey: string,
): Promise<StoredHttpResponse | null> => {
  for (let attempt = 0; attempt < 20; attempt += 1) {
    const record = await store.idempotency.get(route, idempotencyKey);
    if (record?.status === "completed" && record.response !== undefined) {
      return record.response;
    }
    await sleep(25);
  }

  return null;
};

const withIdempotency = async (
  store: ControlPlaneServiceStore,
  route: string,
  idempotencyKey: string,
  requestBody: unknown,
  execute: () => Promise<StoredHttpResponse>,
): Promise<StoredHttpResponse> => {
  const requestHash = hashPayload(requestBody);
  const executeReserved = async (): Promise<StoredHttpResponse> => {
    try {
      const response = await execute();
      await store.idempotency.complete(route, idempotencyKey, response);
      return response;
    } catch (error) {
      await store.idempotency.abandon(route, idempotencyKey);
      throw error;
    }
  };

  const claimOrReplay = async (): Promise<StoredHttpResponse | "pending"> => {
    const reservation = await store.idempotency.reserve({
      routeKey: route,
      idempotencyKey,
      requestHash,
    });

    if (reservation.state === "conflict") {
      throw new Error("Idempotency-Key was reused with a different request body");
    }

    if (reservation.state === "completed") {
      if (reservation.record.response === undefined) {
        throw new Error("completed idempotent response is missing persisted data");
      }
      return reservation.record.response;
    }

    if (reservation.state === "pending") {
      return "pending";
    }

    return executeReserved();
  };

  const initialResult = await claimOrReplay();
  if (initialResult !== "pending") {
    return initialResult;
  }

  const response = await waitForCompletedIdempotency(
    store,
    route,
    idempotencyKey,
  );
  if (response !== null) {
    return response;
  }

  const reacquiredResult = await claimOrReplay();
  if (reacquiredResult !== "pending") {
    return reacquiredResult;
  }

  throw new Error("Idempotent request is already in progress");
};

const assertTenantMatch = (expectedTenantId: string, actualTenantId: string): void => {
  if (expectedTenantId !== actualTenantId) {
    throw new Error("request tenant does not match the authenticated API key");
  }
};

const findSessionForTenant = async (
  repositories: CipRepositories,
  sessionId: string,
  tenantId: string,
) => {
  const session = await repositories.runSessions.getById(sessionId);
  if (session === null) {
    throw new Error(`unknown session ${sessionId}`);
  }
  assertTenantMatch(tenantId, session.tenantId);
  return session;
};

const authenticateSdk = async (
  serviceStore: ControlPlaneServiceStore,
  authorizationHeader: string | undefined,
  requiredScope: ApiKeyScope,
) => {
  const record = await authenticateSdkApiKey(
    serviceStore,
    authorizationHeader,
    requiredScope,
  );
  if (record === null) {
    throw new Error("unauthorized");
  }
  return record;
};

const authenticateOperator = async (
  authorizationHeader: string | undefined,
  operatorAuth: OperatorAuthConfig,
) => {
  const claims = await verifyOperatorToken(authorizationHeader, operatorAuth);
  if (claims === null) {
    throw new Error("unauthorized");
  }
  return claims;
};

const sdkActorFor = (apiKey: ApiKeyRecord) => ({
  type: "agent" as const,
  id: `api-key:${apiKey.id}`,
});

const mergeReportedActor = (
  payload: Record<string, unknown> | undefined,
  reportedActor: unknown,
): Record<string, unknown> =>
  reportedActor === undefined
    ? payload ?? {}
    : {
        ...(payload ?? {}),
        _reportedActor: reportedActor,
      };

const normalizeBatchActors = (
  batch: CipEventBatch,
  apiKey: ApiKeyRecord,
): CipEventBatch => {
  const actor = sdkActorFor(apiKey);
  return {
    tenantId: batch.tenantId,
    sessionId: batch.sessionId,
    events: batch.events.map((event) =>
      event.kind === "run_event"
        ? {
            ...event,
            actor,
            payload: mergeReportedActor(event.payload, event.actor),
          }
        : {
            ...event,
            actor,
            payload: mergeReportedActor(event.payload, event.actor),
          },
    ),
  };
};

const findDeploymentForTenant = async (
  repositories: CipRepositories,
  deploymentId: string,
) => {
  const deployment = await repositories.deployments.getById(deploymentId);
  if (deployment === null) {
    throw new Error(`unknown deployment ${deploymentId}`);
  }
  return deployment;
};

const findApprovalRequestForTenant = async (
  repositories: CipRepositories,
  approvalRequestId: string,
) => {
  const approvalRequest = await repositories.approvalRequests.getById(
    approvalRequestId,
  );
  if (approvalRequest === null) {
    throw new Error(`unknown approval request ${approvalRequestId}`);
  }
  return approvalRequest;
};

const sendError = async (
  error: unknown,
  reply: import("fastify").FastifyReply,
): Promise<void> => {
  if (error instanceof CipControlPlaneError) {
    await reply.code(400).send({ error: error.message });
    return;
  }
  if (error instanceof Error && error.message === "unauthorized") {
    await reply.code(401).send({ error: "unauthorized" });
    return;
  }
  if (error instanceof RequestValidationError) {
    await reply.code(400).send({ error: error.message });
    return;
  }
  if (error instanceof Error && error.message.includes("required scope")) {
    await reply.code(403).send({ error: error.message });
    return;
  }
  if (error instanceof Error && error.message.includes("Idempotency-Key")) {
    await reply.code(400).send({ error: error.message });
    return;
  }
  if (error instanceof Error && error.message.includes("already in progress")) {
    await reply.code(409).send({ error: error.message });
    return;
  }
  if (error instanceof Error && error.message.includes("unknown")) {
    await reply.code(404).send({ error: error.message });
    return;
  }
  if (error instanceof Error && error.message.includes("tenant")) {
    await reply.code(403).send({ error: error.message });
    return;
  }
  await reply.code(500).send({
    error: error instanceof Error ? error.message : "unexpected server error",
  });
};

export const createControlPlaneApiApp = (
  options: ControlPlaneApiAppOptions,
): FastifyInstance => {
  const app = Fastify();

  app.get("/healthz", async (_request, reply) => {
    await reply.code(200).send({ status: "ok" });
  });

  app.get("/readyz", async (_request, reply) => {
    const ready = (await options.readyCheck?.()) ?? true;
    await reply.code(ready ? 200 : 503).send({
      status: ready ? "ready" : "not-ready",
    });
  });

  app.post("/v1/sessions", async (request, reply) => {
    try {
      const apiKey = await authenticateSdk(
        options.serviceStore,
        headerValue(request.headers.authorization),
        "sessions:write",
      );
      const body = parseOrThrow(
        startRunSessionInputSchema,
        request.body,
        "session request",
      ) as StartRunSessionInput;
      assertTenantMatch(apiKey.tenantId, body.tenantId);

      const idempotencyKey = requireIdempotencyKey(
        request.headers as Record<string, unknown>,
      );
      const response = await withIdempotency(
        options.serviceStore,
        routeKey("POST", "/v1/sessions"),
        idempotencyKey,
        body,
        async () => ({
          status: 200,
          body: await options.controlPlane.startRunSession(body),
        }),
      );
      await sendStoredResponse(response, reply);
    } catch (error) {
      await sendError(error, reply);
    }
  });

  app.post(
    "/v1/sessions/:sessionId(^[^:]+)/events::enqueue",
    async (request, reply) => {
    try {
      const apiKey = await authenticateSdk(
        options.serviceStore,
        headerValue(request.headers.authorization),
        "sessions:write",
      );
      const params = parseOrThrow(
        sessionPathSchema,
        request.params,
        "session path",
      );
      const body = parseOrThrow(
        cipEventBatchSchema,
        request.body,
        "event batch",
      ) as CipEventBatch;
      if (body.sessionId !== params.sessionId) {
        throw new Error("request session does not match the route parameter");
      }
      assertTenantMatch(apiKey.tenantId, body.tenantId);
      await findSessionForTenant(options.repositories, body.sessionId, apiKey.tenantId);
      const normalizedBatch = normalizeBatchActors(body, apiKey);

      const idempotencyKey = requireIdempotencyKey(
        request.headers as Record<string, unknown>,
      );
      const response = await withIdempotency(
        options.serviceStore,
        routeKey("POST", "/v1/sessions/:sessionId/events:enqueue"),
        idempotencyKey,
        normalizedBatch,
        async () => {
          const job = createQueuedIngestJob(
            normalizedBatch.tenantId,
            normalizedBatch.sessionId,
            normalizedBatch,
            idempotencyKey,
          );
          await options.serviceStore.ingestJobs.enqueue(job);
          return {
            status: 202,
            body: {
              ingestJobId: job.id,
              acceptedCount: normalizedBatch.events.length,
              receivedAt: job.createdAt,
            },
          };
        },
      );
      await sendStoredResponse(response, reply);
    } catch (error) {
      await sendError(error, reply);
    }
    },
  );

  app.post("/v1/sessions/:sessionId/approval-requests", async (request, reply) => {
    try {
      const apiKey = await authenticateSdk(
        options.serviceStore,
        headerValue(request.headers.authorization),
        "approvals:write",
      );
      const params = parseOrThrow(
        sessionPathSchema,
        request.params,
        "session path",
      );
      await findSessionForTenant(options.repositories, params.sessionId, apiKey.tenantId);
      const body = parseOrThrow(
        requestHumanApprovalInputSchema,
        request.body,
        "approval request",
      ) as RequestHumanApprovalInput;
      if (body.sessionId !== params.sessionId) {
        throw new Error("request session does not match the route parameter");
      }
      const normalizedBody: RequestHumanApprovalInput = {
        ...body,
        actor: sdkActorFor(apiKey),
      };
      await reply.code(200).send(
        await options.controlPlane.requestHumanApproval(normalizedBody),
      );
    } catch (error) {
      await sendError(error, reply);
    }
  });

  app.post(
    "/v1/sessions/:sessionId(^[^:]+)::complete",
    async (request, reply) => {
    try {
      const apiKey = await authenticateSdk(
        options.serviceStore,
        headerValue(request.headers.authorization),
        "sessions:write",
      );
      const params = parseOrThrow(
        sessionPathSchema,
        request.params,
        "session path",
      );
      await findSessionForTenant(options.repositories, params.sessionId, apiKey.tenantId);

      const body = parseOrThrow(
        completeRunSessionInputSchema,
        request.body,
        "complete session request",
      ) as CompleteRunSessionInput;
      if (body.sessionId !== params.sessionId) {
        throw new Error("request session does not match the route parameter");
      }
      const idempotencyKey = requireIdempotencyKey(
        request.headers as Record<string, unknown>,
      );
      const response = await withIdempotency(
        options.serviceStore,
        routeKey("POST", "/v1/sessions/:sessionId:complete"),
        idempotencyKey,
        body,
        async () => ({
          status: 200,
          body: await options.controlPlane.completeRunSession(body),
        }),
      );
      await sendStoredResponse(response, reply);
    } catch (error) {
      await sendError(error, reply);
    }
    },
  );

  app.get("/v1/sessions/:sessionId/replay", async (request, reply) => {
    try {
      const apiKey = await authenticateSdk(
        options.serviceStore,
        headerValue(request.headers.authorization),
        "sessions:read",
      );
      const params = parseOrThrow(
        sessionPathSchema,
        request.params,
        "session path",
      );
      await findSessionForTenant(options.repositories, params.sessionId, apiKey.tenantId);
      await reply.code(200).send(
        await options.controlPlane.replayRunSession(params.sessionId),
      );
    } catch (error) {
      await sendError(error, reply);
    }
  });

  app.get("/v1/evidence-bundles/:sessionId", async (request, reply) => {
    try {
      const apiKey = await authenticateSdk(
        options.serviceStore,
        headerValue(request.headers.authorization),
        "sessions:read",
      );
      const params = parseOrThrow(
        sessionPathSchema,
        request.params,
        "session path",
      );
      await findSessionForTenant(options.repositories, params.sessionId, apiKey.tenantId);
      await reply.code(200).send(
        await options.controlPlane.getEvidenceBundle(params.sessionId),
      );
    } catch (error) {
      await sendError(error, reply);
    }
  });

  app.post(
    "/v1/approval-requests/:approvalRequestId(^[^:]+)::resolve",
    async (request, reply) => {
    try {
      const claims = await authenticateOperator(
        headerValue(request.headers.authorization),
        options.operatorAuth,
      );
      const params = parseOrThrow(
        approvalRequestPathSchema,
        request.params,
        "approval request path",
      );
      const approvalRequest = await findApprovalRequestForTenant(
        options.repositories,
        params.approvalRequestId,
      );
      requireOperatorScope(claims, "approvals:resolve", approvalRequest.tenantId);
      const body = parseOrThrow(
        resolveApprovalRequestInputSchema,
        request.body,
        "approval resolution request",
      ) as ResolveApprovalRequestInput;
      if (body.approvalRequestId !== params.approvalRequestId) {
        throw new Error("request approval id does not match the route parameter");
      }
      const normalizedBody: ResolveApprovalRequestInput = {
        ...body,
        actor: {
          type: "human",
          id: claims.sub,
        },
      };
      await reply.code(200).send(
        await options.controlPlane.resolveApprovalRequest(normalizedBody),
      );
    } catch (error) {
      await sendError(error, reply);
    }
    },
  );

  app.post(
    "/v1/deployments/:deploymentId(^[^:]+)::transition",
    async (request, reply) => {
    try {
      const claims = await authenticateOperator(
        headerValue(request.headers.authorization),
        options.operatorAuth,
      );
      const params = parseOrThrow(
        deploymentPathSchema,
        request.params,
        "deployment path",
      );
      const deployment = await findDeploymentForTenant(
        options.repositories,
        params.deploymentId,
      );
      requireOperatorScope(claims, "deployments:write", deployment.tenantId);
      const body = parseOrThrow(
        transitionDeploymentInputSchema,
        request.body,
        "deployment transition request",
      ) as TransitionDeploymentInput;
      if (body.deploymentId !== params.deploymentId) {
        throw new Error("request deployment id does not match the route parameter");
      }
      const normalizedBody: TransitionDeploymentInput = {
        ...body,
        actor: {
          type: "human",
          id: claims.sub,
        },
      };
      await reply.code(200).send(
        await options.controlPlane.transitionDeployment(normalizedBody),
      );
    } catch (error) {
      await sendError(error, reply);
    }
    },
  );

  app.post(
    "/v1/deployments/:deploymentId(^[^:]+)::rollback",
    async (request, reply) => {
      try {
        const claims = await authenticateOperator(
          headerValue(request.headers.authorization),
          options.operatorAuth,
        );
        const params = parseOrThrow(
          deploymentPathSchema,
          request.params,
          "deployment path",
        );
        const deployment = await findDeploymentForTenant(
          options.repositories,
          params.deploymentId,
        );
        requireOperatorScope(claims, "deployments:write", deployment.tenantId);
        const body = parseOrThrow(
          rollbackDeploymentInputSchema,
          request.body,
          "deployment rollback request",
        ) as RollbackDeploymentInput;
        if (body.deploymentId !== params.deploymentId) {
          throw new Error("request deployment id does not match the route parameter");
        }
        const normalizedBody: RollbackDeploymentInput = {
          ...body,
          actor: {
            type: "human",
            id: claims.sub,
          },
        };
        await reply.code(200).send(
          await options.controlPlane.rollbackDeploymentToBlueprint(normalizedBody),
        );
      } catch (error) {
        await sendError(error, reply);
      }
    },
  );

  app.get("/v1/deployments", async (request, reply) => {
    try {
      const claims = await authenticateOperator(
        headerValue(request.headers.authorization),
        options.operatorAuth,
      );
      const query = parseOrThrow(
        optionalTenantQuerySchema,
        request.query,
        "deployment query",
      );
      const tenantId = query.tenantId ?? claims.tenantId;
      requireOperatorScope(claims, "deployments:read", tenantId);
      await reply.code(200).send(
        await options.repositories.deployments.list(
          tenantId === undefined ? undefined : { tenantId },
        ),
      );
    } catch (error) {
      await sendError(error, reply);
    }
  });

  app.get("/v1/tenants/:tenantId", async (request, reply) => {
    try {
      const claims = await authenticateOperator(
        headerValue(request.headers.authorization),
        options.operatorAuth,
      );
      const params = parseOrThrow(
        tenantPathSchema,
        request.params,
        "tenant path",
      );
      requireOperatorScope(claims, "tenants:read", params.tenantId);
      await reply.code(200).send(
        await options.repositories.tenants.getById(params.tenantId),
      );
    } catch (error) {
      await sendError(error, reply);
    }
  });

  app.get("/v1/audit-events", async (request, reply) => {
    try {
      const claims = await authenticateOperator(
        headerValue(request.headers.authorization),
        options.operatorAuth,
      );
      const query = parseOrThrow(
        optionalTenantQuerySchema,
        request.query,
        "audit query",
      );
      const tenantId = query.tenantId ?? claims.tenantId;
      requireOperatorScope(claims, "audit:read", tenantId);
      await reply.code(200).send(
        await options.repositories.auditEvents.list(
          tenantId === undefined ? undefined : { tenantId },
        ),
      );
    } catch (error) {
      await sendError(error, reply);
    }
  });

  return app;
};